requirements.txt
streamlit
pypdf
python-docx
langchain
langchain-huggingface
langchain-groq
langchain-community
groq
faiss-cpu
sentence-transformers
